Add an option to limit the number of checkpoints
Summary:
While training the NMT, we write a bunch of checkpoints. This patch adds an
option `--max-checkpoints-kept` that allows to specify how many checkpoints
are kept alive.

When this option is specified, only the N last checkpoints (and N last average
checkpoints) are kept around.

The default behavior matches what the program was doing before this patch,
i.e., it keeps all checkpoints around.

Differential Revision: D7867496

fbshipit-source-id: c1facc183f96df579b72fdc13f8718e48fa95351