Use binarized data format instead of raw text for memory/reuse efficiency
Summary:
If the user passes in a text file, it's binarized and the rest of training code uses the binarized version.

The binarized file is either saved to the location specified by the flag (for re-use) or to a temp file.

Add bool_flag helper type to allow flags like --reverse-source (which default to true) to be turned off by specifying "--reverse-source False"

Depends on https://github.com/pytorch/fairseq/pull/161

Closes https://github.com/pytorch/translate/pull/62

Reviewed By: liezl200

Differential Revision: D8021994

fbshipit-source-id: c9ed17a17f448d9ad206580d8aa438fafb9ff941