char_rnn architecture
Summary:
Closes https://github.com/pytorch/translate/pull/80

Introduces data loading which encodes source sentences at both the token level (in the same way as the default architecture) and the characters per token level. Numberized indices for both are inputs to the encoder.

Also introduces CharSourceModel (--arch char_source), which consumes input in this format. This model initially supports a CharRNNEncoder, replacing the existing CharRNN module, which required text input in a special format (whitespace separated characters, with a special token delimiting words). This model will allow for faster training of character RNN model, as well as using the existing vocabulary reduction mechanism (and, in a future diff, a hybrid word/character encoder).

Reviewed By: theweiho, xianxl

Differential Revision: D8181361

fbshipit-source-id: 07c2937ceed8a4d9866ac5a9ebc0aa1291c79926