squash .data
Summary:
Quick simplification of the code base removing `.data` annotations on tensors. This is generally unnecessary since the merging of tensors and variables.

Exceptions are when calling an in-place operation on a leaf variable (here meaning parameter, i.e., tensor which is not computed and requires grad), such as:

  param.data.uniform_(-0.1, 0.1)

And when calling `.numpy()` on a tensor which requires grad

Differential Revision: D8208506

fbshipit-source-id: b9622bff8b9aafb0d9522675af5d1917290613de