squash .data
Summary:
Closes https://github.com/pytorch/translate/pull/82

Quick simplification of the code base removing `.data` annotations on tensors. This is generally unnecessary since the merging of tensors and variables.

Exceptions are when calling an in-place operation on a leaf variable (here meaning parameter, i.e., tensor which is not computed and requires grad), such as:

  param.data.uniform_(-0.1, 0.1)

And when calling `.numpy()` on a tensor which requires grad

Reviewed By: liezl200

Differential Revision: D8208506

fbshipit-source-id: 1c7d967a9480aac190a6040611e93f1a2f74a996