Fixed bug where `generate_score` crashes when invoked from train.py
Summary: In a previous version, the procedure to evaluate BLEU score in `train.py` invokes `_generate_score` in `generate.py`. This tries to save translations if the `translation_output_file` arg is set. This argument is not (and probably should not be) an argument in `train.py`. Changed it to assume False when this argument is not present, since we don't want to save translations during training.

Reviewed By: liezl200, pmichel31415

Differential Revision: D8230928

fbshipit-source-id: b0485a9ef8627b1f9aba0f7b410984153c5266fd