remove default arguments for char_embed_dim, token_embed_dim
Summary: Rectifying inconsistency between command-line default and constructor default. There's probably not a good reason to have default values for the constructor arguments at all, but I've erred on the side of leaving it as it is except for the command-line values whose defaults appear immediately above in the same file.

Differential Revision: D8283159

fbshipit-source-id: 52093d23d903ab528622f407c2140178c6946463