remove default arguments for char_embed_dim, token_embed_dim (#89)
Summary:
Closes https://github.com/pytorch/translate/pull/89

Rectifying inconsistency between command-line default and constructor default. There's probably not a good reason to have default values for the constructor arguments at all, but I've erred on the side of leaving it as it is except for the command-line values whose defaults appear immediately above in the same file.

Reviewed By: xianxl

Differential Revision: D8283159

fbshipit-source-id: e990b4f1dadaf59c163781e67d24f9d4f4ee16d1