Multi-decoder adaptive ensembles (#98)
Summary:
Closes https://github.com/pytorch/translate/pull/98

This diff implements adaptive ensembles. We train multiple decoder networks, and combine their predictions at the end. All
decoders are jointly trained and share the encoder and potentially the output projection layer. The way the predictions are
combined can be controlled with --multi-decoder-combination-strategy.

This diff is inspired by D8157064.

Reviewed By: jhcross

Differential Revision: D8359267

fbshipit-source-id: fa322b03a4ac0597a4e8d7e57a5b9d7619af5c51