forced decoder network (onnx export) (#95)
Summary:
Closes https://github.com/pytorch/translate/pull/95

Add networks in ensemble_export.py to create a forced decoding network from PyTorch NMT checkpoints. This network takes an arbitrary numberized (source, target) pair and returns the model score for the translation, including penalties.

Vocabulary reduction networks are also supported, but note that target indices which are not in the possible_translation_tokens generated for the source input will be treated as UNK.

Reviewed By: jamesr66a

Differential Revision: D8336745

fbshipit-source-id: 6b8de5014adbb4eca080804de87d474c98af9917