ONNX export for Character-RNN (#101)
Summary:
Closes https://github.com/pytorch/translate/pull/101

Enables ONNX export for CharRNNEncoder by short-circuiting the code path with unsupported operations (sort) and tracing an already-sorted (or rather uniform-word-length) example in PyTorch. This should work because ONNX/Caffe2 do not have the constraint that a batch of sequence input to LSTM be sorted in descending order of length.

Reviewed By: xianxl

Differential Revision: D8371501

fbshipit-source-id: 5dbad52bc60e7f06a7928c5cf217d273f51570f9