ONNX export for Character-RNN (#111)
Summary:
Closes https://github.com/pytorch/translate/pull/111

Closes https://github.com/pytorch/translate/pull/101

Enables ONNX export for CharRNNEncoder by short-circuiting the code path with unsupported operations (sort) and tracing an already-sorted (or rather uniform-word-length) example in PyTorch. This should work because ONNX/Caffe2 do not have the constraint that a batch of sequence input to LSTM be sorted in descending order of length.

Reviewed By: xianxl

Differential Revision: D8371501

fbshipit-source-id: 857e3a8c8d58c5054ce2e3c14a4d771f4c3a9af9