Component-based Multilingual NMT (#93)
Summary:
Closes https://github.com/pytorch/translate/pull/93

This diff makes it possible to jointly train several language pairs. We train separate encoder networks for each source language, and separate decoder networks for each target language. During training, sentences from any corpus (and language pair) are mixed together in a batch, and encoder and decoder networks are selected based on the language IDs.

Potential applications of this diff are:
- Improving NMT in general by sharing encoder and decoder networks across language pairs
- Improving low-resource NMT in particular as encoder and decoder networks by their own receive more training data when shared
- Zero-shot translation
- Integrating monolingual training data with multiling-source-lang = multiling-target-lang

Reviewed By: jhcross

Differential Revision: D8187615

fbshipit-source-id: 30714cd9429dfe5c0a767b040d9709ffc85d3ff8