Component-based Multilingual NMT (#94)
Summary:
Closes https://github.com/pytorch/translate/pull/94

This diff makes it possible to jointly train several language pairs. We train separate encoder networks for each source language, and separate decoder networks for each target language. During training, sentences from any corpus (and language pair) are mixed together in a batch, and encoder and decoder networks are selected based on the language IDs.

This is part two of D8187615 containing the multilingual encoder and decoder implementations, and multilingual beam search.

Reviewed By: jhcross

Differential Revision: D8333915

fbshipit-source-id: c7e6d3dafbb272527f318691920d96e747d308c9