Add the option to load pretrained weights for training (#116)
Summary:
Closes https://github.com/pytorch/translate/pull/116

This lets the user specify a checkpoint from which the weights of the models will be loaded before training (but the optimizer will still be reset).

An example where this is useful is to continue training an existing model with a different optimizer (eg fine-tuning with SGD + small learning rate for domain adaptation).

Reviewed By: dpacgopinath

Differential Revision: D8465146

fbshipit-source-id: 157a63a0569248a6bae5f0a0a15932ca73514c18