Tie decoder embeddings (#117)
Summary:
Closes https://github.com/pytorch/translate/pull/117

Makes it possible to tie the decoder word embeddings and the softmax weights. From my experience this is often helpful especially for small datasets (also this takes less parameters).

Here is a reference for MT (this has been introduced in LM before): https://arxiv.org/abs/1710.01329 (NAACL18)
 ---

Here are two experiments with the same configuration except for untied/tied embeddings. This was on the iwslt data provided on the pytorch_translate github (de-en). Refer to the chronos links for full model details.

- base (untied): https://fburl.com/chronos/tjf04wnr
  - best valid BLEU: 31.30
  - checkpoint size on disk: 453M
- tied: https://fburl.com/chronos/j3rajgvw
  - best valid BLEU: 31.61
  - checkpoint size on disk: 374M

- Small BLEU score gain (+0.31)
- reduces parameter cost by ~17%

Reviewed By: jmp84

Differential Revision: D8465257

fbshipit-source-id: 27ce7b52ab5753067814bdfc9204b2c43718a484