Introduce NoAttention type
Summary:
RNNDecoder currently allows attention_type to be None, but setting it to None is not accessible
via command-line, and setting it programmatically crashes.

Introduces a new NoAttention attention type, which is a dummy implementation of `BaseAttention`
for not using any attention. This attention type replaces the attention_type=None use case from before.

We also rename AttentionType.encoder_output_dim to AttentionType.context_dim since it is set to 0 for NoAttention.

Reviewed By: yukatherin

Differential Revision: D8654417

fbshipit-source-id: 1ce73eff562a209900be6a1fc41260bcff0e640a