Introduce NoAttention type (#140)
Summary:
Closes https://github.com/pytorch/translate/pull/140

RNNDecoder currently allows attention_type to be None, but setting it to None is not accessible
via command-line, and setting it programmatically crashes.

Introduces a new NoAttention attention type, which is a dummy implementation of `BaseAttention`
for not using any attention. This attention type replaces the attention_type=None use case from before.

We also rename AttentionType.encoder_output_dim to AttentionType.context_dim since it is set to 0 for NoAttention.

(Note: this ignores all push blocking failures!)

Reviewed By: yukatherin

Differential Revision: D8654417

fbshipit-source-id: 3678be615abd92ce66767c4f1ce6781a79e26ba9