Interface for pruning (#151)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/151

Adds an optional step to base_training which prunes a fixed proportion of connections and then trains while fixing those connections at zero. This implements class-blind pruning from (See, Luong, Manning 2016).

For the sake of simplicity and consistency with the finetuning code, this commit removes support for train-prune-retrain in pytorch translate.

Reviewed By: dpacgopinath

Differential Revision: D8772690

fbshipit-source-id: 4c785d043242ae4a0b0e5988438afba67e41f3ea