Beam search which keeps completed hypos in the beam (#161)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/161

This is a beam search implementation in PyTorch which keeps completed hypos in the beam.
Completed hypos are passed through the next time step, and keep competing against hypo
expansions. Beam search terminates if all active hypos end with EOS.

Leaving this under research/ because experimentation did not show a noticeable difference in translation quality from the existing implementation.

Reviewed By: jhcross

Differential Revision: D8717422

fbshipit-source-id: 577b0d582ecb9775238420377b7ad1e8282079c3