modularize words prediction
Summary:
A series of diffs to extend words prediction so
that it can be used as vocab reduction purpose on any arch.

The code under research/words_prediction has lots of reduandant logic, both in
the loss and model arch. This diff tries to make words prediction a plug and
play which can work with existing translation loss and encoder/decoder classes.

Reviewed By: jhcross

Differential Revision: D9012503

fbshipit-source-id: 6e0edf874c7c026178af101acfacb35285aeedbc