Add adversarial criteria
Summary:
This adds two criterions geared towards adversarial attacks:

- `all_bad_words`: log probability of **not** generating the correct word at each timestep. This is different from `cross_entropy --reverse-criterion`. The latter can be seen as maximizing `P[NOT (w1 AND w2 ...)]=P[(NOT w1) OR (NOT w2) ...]` whereas this criterion maximizes P[(NOT w1) AND (NOT w2) ...]
- `force_words`: log probability of generating a specific set of words (eg. force the model to generate profanities).
- `force_words_hinge`: Same as `force_words` but with a hinge-like loss inspired by https://arxiv.org/abs/1803.01128
 ---

All experiments are run on the first 1000 samples of IWSLT de-en. The complete details of the model can be found below. I used greedy decoding with a length penalty of 1.0.

In this first experiment I show that this criterion is better than the reversed NLL criterion (`cross_entropy --reverse-criterion`). All experiments are run with `--max-swaps=1`.

Here are the BLEU scores. Lower BLEU = bigger degradation = better.

|Adversary\Criterion|`cross_entropy` (reversed)|`all_bad_words`|
|--|
|Baseline (no adversary)|24.01|24.01|
|`random_swap`|**19.14**|19.14|
|`random_swap --by-gradient-norm --temperature 10`|19.17|19.13|
|`brute_force`|20.75|**18.67**|
|`brute_force --normalize-directions`|20.94|20.90|

This criterion is better for gradient based attacks

I now show that the `force_words` criterion can be used to force the model to generate specific words. I force the model to generate words from the following list:

`apfel banane birne ananas orange` (`apple banana pear ananas orange`)

We then count the number of occurrences of the target words in the generated sentences.

This attack works better when the original target sentence is replaced with the sentence generated by the model (because we want to deviate from the model output)

|Adversary\counts|# apfel |# banane |# birne |# ananas |# orange|# all|
|--|
|Baseline (no adversary)|0|0|0|0|0|0|
|`random_swap`|0|0|0|0|0|0|
|`random_swap --by-gradient-norm --temperature 10`|0|0|0|0|0|0|
|`brute_force`|1|0|0|0|1|2|
|`brute_force --normalize-directions`|**6**|0|0|0|**5**|**11**|

It is possible (albeit difficult) to force the model to generate specific words. In this case, a gadient based brute force attack (*with normalized directions* works better).

The setting is the same as experiment 2. except the adversary is now fixed (`brute_force --normalize-directions`) and I vary the criteria used.

|Criterion\counts|# apfel |# banane |# birne |# ananas |# orange|# all|
|--|
|`force_words`|6|0|0|0|5|11|
|`force_words --any-position`|**20**|0|0|0|**7**|**27**|
|`force_words_hinge`|3|0|0|0|1|4|
|`force_words_hinge --any-position`|4|0|0|0|1|5|

The probabilistic approach (`force_words`) with logsumexp reduction (`--any-position`) works much better than the other options. In particular this is an improvement over the hinge-like loss proposed in https://arxiv.org/abs/1803.01128 (however their search for adversarial examples is different so the results are not directly comparable).

With the following criterion: `force_words --force-not --words-list </s>` I am able to decrease the BLEU score to 17.53 which is worse than every other approach in experiment 1.

More precisely the length ration hyp/ref goes from 1.126 to 1.401. Although this doesn't change the brevity pernalty (capped to 1.0 anyway), this decreases the precisions a lot.

This experiment was still conducted with a beam size of 1. For a beam size of 5 the BLEU score degrades from 25.21 to 20.31 (ratio 1.091 -> 1.267). However in this setting random swapping results in a similar if not better degradation in BLEU (but not in length ratio).

Reviewed By: xianxl

Differential Revision: D8624012

fbshipit-source-id: 94fdd4fe478893024576430aa2138f26881a7650