Make installation instruction less scary
Summary:
Right now the github installation instruction require the user to go through an obstacle course including (but not limited to)

- Uninstalling pytorch *twice*
- Installing caffe2
- Reinstall numpy
- Build the C++ decoding code (with cmake in bonus)

This is enough to discourage most potential users.

The bulk of the complexity here comes from the ONNX/Caffe2 part which is not needed for the majority of use cases.

This adds a Quickstart section with simplified installation instructions.

Created from Diffusion's 'Open in Editor' feature.

Reviewed By: theweiho, jhcross

Differential Revision: D9212303

fbshipit-source-id: 377c696f0b923a4fc301d40071553e8e68ef8295