maybe_cuda to allow reordering of DummyEncoder output (#203)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/203

To enable beam search (though its utility is questionable) we need to allow `reorder_encoder_out()` to operate correctly on the DummyEncoder (effectively just scaling up the size of the batch dimension by beam_size). To do this, the tensors need to be on GPU (when using CUDA).

Note: it doesn't really make sense to be doing BLEU evals when training a language model anyway, so in a future diff, we should simply gate the evals in train.py so that they are not used for LM-only training.

Reviewed By: lhu17

Differential Revision: D9375692

fbshipit-source-id: c1698ec3be794fd7c79d0ea6c2de2180a1cfee47