Ensembling support in generation for models with different output softmax size (#210)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/210

In general, this diff enables the ensembling of multiple models with different sizes of output softmax (for example where model 1 outputs a set of 10 translation candidates + 10 probabilities for each of those candidates, and model 2 outputs a set of 20 translation candidates + 20 probabilities for each of those candidates). See documentation in comments for a full example.

The context of this diff is to allow ensembling of multiple models trained with word prediction vocab reduction.

Reviewed By: xianxl

Differential Revision: D9504725

fbshipit-source-id: 2c5747ce9d7a1bfd556fe3d9362a53e3353b43ee