predict top-k labels based on number of source tokens
Summary: When doing predictive vocabulary reduction, it makes sense to determine how many labels to select based on the number of source tokens. This prevents us wasting space and time when we have large batches of short sequences. Implemented as a simple integer multiple of number of source tokens, provided as `--topk-labels-per-source-token`.

Reviewed By: jmp84

Differential Revision: D9784727

fbshipit-source-id: 56b7b5765bca7f1376e5cc139ce4a52b88429aef