Fix MI-LSTM GPU training extended test failure
Summary:
There were 2 test failures. This is the one that happens on GPU
machines. Will look at the other test failure independently.

There were 2 issues in this test. One is not calling the parent's init with
the correct parameters. And the other one was fairseq trainer requiring
all gradient receiving parameters to be in the forward pass computation.
(https://fburl.com/g6xtyh80)

Reviewed By: jhcross

Differential Revision: D9990597

fbshipit-source-id: 1b6c94e609502f68e28bd211f95bcb9535f1c40d