Remove `SerialOperatorExecution` from dual learning workflow
Summary:
This is blocking D8987015, because we couldn't specify entitlement together with `SerialOperatorExecution` (T33242195).

Phabricator diff view is not great for simple indent changes so here are the actual changes:  remove the `SerialOperatorExecution` statements, reindent and move the copy checkpoint logic to an async (otherwise we get a future related error).

Reviewed By: xianxl

Differential Revision: D9953881

fbshipit-source-id: 94370bf4b822223a0f38b5c0a713db054c037eb0