net surgery: fuse Transpose and BatchMatMul (#241)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/241

This change to ONNX export speeds up inference for transformer networks by taking advantage of the the `trans_a` and `trans_b` arguments for the Caffe2 BatchMatMul operator. It an input to BatchMatMul is the output of a Transpose operator (and no other operator), then BatchMatMul can "implicitly" do the transpose at no extra computational cost by supplying those arguments.

When exporting, we identify this situation, and remove the unnecessary Transpose op, using the BatchMatMul arguments instead.

Reviewed By: jspark1105

Differential Revision: D10387151

fbshipit-source-id: 9a9bf4865fc9ee0ffdfe7530d994511497928778