generate adversarial examples at embeddings level
Summary:
Create a new adversarial trainer that perturbs token embeddings in a way that most greatly decreases a training criterion (according to one gradient step). See https://arxiv.org/pdf/1605.07725.pdf.

This is implemented by modifying `adversarial_trainer.py` to calculate the criterion's gradient wrt to token embeddings, and adding a trainer that can set the encoder's `embed_noising_layer` to modify embeddings using this gradient.

Also, add flag for normalizing RNN encoder embeddings to have zero mean and unit variance (when weighted by token frequency). This is useful for the adversarial training so that the perturbation norm will be set relative to the embedding norm.

Also add configs for fblearner and a unittest.

Reviewed By: xianxl

Differential Revision: D12851437

fbshipit-source-id: 84e4733dc2211b6384dd6662242e01e32d997c3b