Set use_output_queue=True
Summary: I was able to backtrace some of our reproducibility issues to this flag, because when we set this to False (unlike base_training.py https://fburl.com/1mxg1k93), we end up not syncing data (https://fburl.com/hulxi629). So let's make sure, train.par is as close as to production possible.

Reviewed By: theweiho

Differential Revision: D13198658

fbshipit-source-id: 3c77ca768849c6c38567e1a60d6af936cc5ffc3a