Resend D13433333 without the side effect
Summary:
I previously sent D13433333 for refactoring with a single side effect. I stopped sending train_data (that we use to plot graphs in fblearner) just for end of epoch. I thought we are sending it at every iteration already, so not sending at the end of epoch would be a trivial decision and graphs would be fine.

But, I was wrong because there was an assertion (without a test) checking this. So, that assertion caused our runs to fail in prod.

So, for now I am resending D13433333 without the side effect, but I think we should have unit tests for the assertions or assumptions we have in base_training. We had a lot of prod failures last half that we should catch with unit tests, therefore created a task for it T38768352 and will work on that

Reviewed By: jhcross

Differential Revision: D13581224

fbshipit-source-id: 6ebbb49667490d386dd91d147a4c63fcb84571b2