Remove expensive calculate_bleu_on_subset operation at the beginning of training
Summary:
One of the top reasons for random distributed failure is "Wait timeout for name(s): 0_0", which means some processes are waiting for rank 0 to finish some work.

We have this calculate_bleu_on_subset, which runs before training in case of a previous checkpoint. It takes quite a while for high resource languages and causes random failures in the system.

Normally, we have 30mins timeout for a process to wait for another process during training. However, before the training starts, somehow, this timeout is not respected.
Pieter Noordhuis mentioned that there seems to be an underlying issue with NCCL backend lazy initialization which won't initialize until it is needed, which is during training. I tried putting a barrier but it doesn't help either. (Plus, we can't provide custom timeout for the barrier)

For now, the best solution seems to be removing this expensive operation that happens before training that skews the timeline.

Reviewed By: jhcross

Differential Revision: D13608751

fbshipit-source-id: 4f6ac6ed41164ff0f47299ec105684afb49d9927