forward-backward algorithm
Summary:
This diff aims to implement the forward-backward algorithm for the segment-HMM model. An implmentation of its vanilla model can be found in: http://www.katrinerk.com/courses/python-worksheets/demo-the-forward-backward-algorithm
Ours is different because ours is based on segment-HMM not the regular HMM.

The next diff should use the forward-backward algorithm to calculate expected values of variables in the EM algorithm.

Differential Revision: D13788034

fbshipit-source-id: 9e3340965f2bb87716d3c11d21394a24c116d464