forward-backward algorithm (#322)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/322

This diff aims to implement the forward-backward algorithm for the segment-HMM model. An implmentation of its vanilla model can be found in: http://www.katrinerk.com/courses/python-worksheets/demo-the-forward-backward-algorithm
Ours is different because ours is based on segment-HMM not the regular HMM.

The next diff should use the forward-backward algorithm to calculate expected values of variables in the EM algorithm.

Reviewed By: akinh

Differential Revision: D13788034

fbshipit-source-id: 90d52b23f294e20704dd4793952cdda6774686de