Using top-k in word-level knowledge distillation (#341)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/341

In word-level knowledge distillation, we should have an option to use the top-k words from the teacher model in the KD loss function. This diff does this. For backward compatibility, we assume that if we set top-k=0, the model backs up to the original approach of evaluating the teacher probability of all words in vocabulary.

Reviewed By: jhcross

Differential Revision: D13994805

fbshipit-source-id: e0d88af957dbdcc730bdfec2f3fcc960b871237e