Morphology initialization with random sampling from a normal distribution (#337)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/337

We are trying to test if random initialization of morpheme distributions is better for unsupervised morphology. This can also be helpful for cases such as the Viterbi-EM (aka hard EM) algorithm (not implemented yet). Here we initialize a segmentation per word by assuming that a prefix/suffix length follows a normal distribution (default: mean: 2 characters, std_dev: 1 character). For initialization we let a word to have at most two prefixes and two suffixes. We also let a word to have two stems as initialization (we break one stem into two parts evenly).

Reviewed By: liezl200

Differential Revision: D14017270

fbshipit-source-id: 8324262163b47b7c2f4fbf665906ea1c5d1b68d5