Fix incremental state bug with hybrid transformers
Summary:
Hybrid transformer was outputting lower scores when using incremental states. Debugging the issue, we realized reordering doesn't work at all, so the scores are not computed with the correct states of the beams.

This diff fixes the issue by reordering the state correctly, in case of order change in beam.

Differential Revision: D14195391

fbshipit-source-id: 50ae9bc2471eedfc5d63913d9c0eaadb08d22107