Fix incremental state bug with hybrid transformers (#353)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/353

Hybrid transformer was outputting lower scores when using incremental states. Debugging the issue, we realized reordering doesn't work at all, so the scores are not computed with the correct states of the beams.

This diff fixes the issue by reordering the state correctly, in case of order change in beam.

Reviewed By: jhcross

Differential Revision: D14195391

fbshipit-source-id: bc3a50ea13e1f20af41aef988ab3dd7dee6737ea