Parameter holder for the bilingual model
Summary:
In this commit, I added a parameter holder with its corresponding unit tests for bilingual morphology.

In standard morphology, the probability of a word is equal to probability of its corresponding morphemes:
$$p(w=[m_1,...,m_k]) = \prod_{i=1}^{k} p(m_i)$$

Here we add an alignment link to the target language (similar to IBM model 1):
$$p(w=[m_1,...,m_k]) = \prod_{i=1}^{k} p(m_i) \sum_{j=1}^{n} t(f_j | m_i)$$

where f_j is the jth morpheme in the target sentence.

Some info on IBM model 1 (note that the vanilla IBM model 1 is word-based but here we are going to deal with morphemes):
http://www.cs.columbia.edu/~mcollins/courses/nlp2011/notes/ibm12.pdf

Differential Revision: D14333029

fbshipit-source-id: 04251c86f907d19f767de5c3d6db36c20eed5644