Adding Viterbi segmentation with respect to translation probabilities (#375)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/375

Here I added a segmentation model that in addition to searching over all morphemes in the source, it also maximizes for the best morpheme in the target language. This is beneficial for Viterbi-EM (future diffs).

$$p(w=[m_1,...,m_k]) = \arg\max \prod_{i=1}^{k} p(m_i) \arg\max_{j=1}^{n} t(f_j | m_i)$$

where f_j is the jth morpheme in the target sentence.
Some info on IBM model 1 (note that the vanilla IBM model 1 is word-based but here we are going to deal with morphemes):
http://www.cs.columbia.edu/~mcollins/courses/nlp2011/notes/ibm12.pdf

Reviewed By: akinh

Differential Revision: D14353265

fbshipit-source-id: 2831db4f4e8d5190396aa89d48b724539b6822b8