Use int64 to store offsets in datasets
Summary:
Due to overflow of numpy int32 because of large number of words in multilingual runs, `offsets` list which stores the indices of beginning of sentences became negative. The batch sampler used these negative numbers as negative indices and produced corrupted batches. As a result, training failed since the final token of (corrupted) sentences was not eos.

See example failure in f104704295

Differential Revision: D14672537

fbshipit-source-id: 20fa498e09b44aed8a5dd853a9864584ff9307b5