Use int64 to store offsets in datasets (#411)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/411

Due to overflow of numpy int32 because of large number of words in multilingual runs, `offsets` list which stores the indices of beginning of sentences became negative. The batch sampler used these negative numbers as negative indices and produced corrupted batches. As a result, training failed since the final token of (corrupted) sentences was not eos.

Reviewed By: jmp84

Differential Revision: D14672537

fbshipit-source-id: b09c9eefc862fedd45ef2ec6defa300fccbbdd97