Fast BPE implementation with memoization (#491)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/491

In this diff, we go beyond the original BPE implementation and use bookkeeping to keep track of candidates that need to change. In this way, we do not need to rebuild candidate frequencies as well as not needing to try to merge everything including those words where a candidate does not exist.

Differential Revision: D15033109

fbshipit-source-id: b0543d1c562a0af243566b015f2faf5b56514e6c