--output-source-binary-path option (#531)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/531

The `--max-examples-to-evaluate` option allows random filtering of a loaded dataset, which can then be used as sequence-level knowledge distillation targets by saving the output hypotheses (oracle outputs from beam if using `--nbest`) with the `--output-hypos-binary-path` flag. To use these sequences for training, however, we also need to save the filtered source, which this change enables (using the `--output-source-binary-path` option.

Differential Revision: D15323145

fbshipit-source-id: b5b296f212509e3730926a4d5a4e6f0d4e700c83