custom LanguagePairUpsamplingDataset
Summary: Extends from fairseq ConcatDataset. As Myle pointed out in D15720559, ConcatDataset is not designed to wrap LanguagePairDataset so in our case there are some limitations. Create a custom dataset for doing it. If we observe serious performance issue for iterating over sizes we could switch back to only supporting integer as sampling_ratios

Differential Revision: D15730883

fbshipit-source-id: 646d6fccf9c5b37a7c8f4936648e6d9d7e3eaf51