custom LanguagePairUpsamplingDataset (#571)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/571

Extends from fairseq ConcatDataset. As Myle pointed out in D15720559, ConcatDataset is not designed to wrap LanguagePairDataset so in our case there are some limitations. Create a custom dataset for doing it. If we observe serious performance issue for iterating over sizes we could switch back to only supporting integer as sampling_ratios

Reviewed By: xianxl

Differential Revision: D15730883

fbshipit-source-id: e4aa019435db566ebdd27103ac433a66805789c7