Fix public training script train_iwslt14.sh by adding an __init__.py
Summary:
The public scripts we have to introduce people to pytorch/translate are broken, ever since we added `pytorch_translate.research.rescore` to the list of imports for `train.py`. This fixes the `ModuleNotFound` error that arises running the script; this one folder didn't have an `__init.py__`, so things broke. All I've done is create that empty file, so that `setuptools.find_packages()` can find the folder.

In general, I think it's jarring for new users of pytorch/translate to see the training script cluttered with imports they don't use. I know it's helpful for us because then everything gets registered, but hopefully we can brainstorm a better way to do this moving forward.

Differential Revision: D16023002

fbshipit-source-id: cc3b62ac5c3931280758d0181d5b315fbec9a877