Character Aware Decoder (Part 1) (#624)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/624

A modified version of the word-based decoder but in here we get the character CNN embedding of the generated tokens.

It is worth mentioning that we are not evaluating open vocabularies, so we can precompute the character CNN representation for every word offline, and then update the word embeddings, save them, and use a vanilla word-based decoder. Later diffs should address these issues.

Reviewed By: jhcross

Differential Revision: D16787078

fbshipit-source-id: 9baef41486da454ea2322f1c3b9152b98b4f87b4