Unit test for char encoder (#634)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/634

Added unit test for char encoder proving that by correct padding we can get the same results of the same inputs in different minibatches.

Context:
It looks like we need more padding for the right side (at least equal to the kernel size). Since the padding is two-sided, what I could do was to double the padding size in order to make sure that in principle we do the correct padding.

Differential Revision: D16995372

fbshipit-source-id: 65f66fcb6d5682d7946aef20dce16ae7b583f6dd