Add back the logic to output translation scores (#653)
Summary:
Pull Request resolved: https://github.com/pytorch/translate/pull/653

This logic was accidentally removed in D15171344. Need to find a mechanism to prevent this in the future. Admittedly, how to come up with a test to catch all zeros score (the consequence of the change by D15171344) vs. the scores were actually from the model?
On the other hand, documentation could potentially help when deleting code one is not familiar with. So, I added extra documentation to reiterate this logic is in use.

Reviewed By: akinh

Differential Revision: D17961721

fbshipit-source-id: df44d3bfa851f338021e4322f8ab3d933f764de7