Add a test for fp16 breakage
Summary:
Our master is failing with a diff from 10 days ago and I realized we don't have a test to alert us about the failure. Let's add a test that shows the breakage before actually solving it.

I was first gonna add an E2E test, but then realized, we probably shouldn't, since this is inside training. All we need is an integration test.

theweiho I agree with your assesment that we don't have to have E2E multi-gpu tests, if we can make these integration tests work on Sandcastle.

Reviewed By: cndn

Differential Revision: D21312488

fbshipit-source-id: 997a26edc93f934ec996c019e2d140188e5a9844